Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
PyJWT==2.9.0
bcrypt==4.2.0
email-validator==2.2.0
psycopg2-binary==2.9.9
asgiref==3.8.1
